requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pandas>=2.2.2
python-docx>=1.1.2
pdfplumber>=0.11.0
python-pptx>=0.6.23
openpyxl>=3.1.3
PyYAML>=6.0.1
regex>=2024.5.15
tiktoken>=0.7.0
